♻️ refactor:  WidgetのComponent化
#92
